+++ neuron/src/app/Neuron/CLI/Query.hs
@@ -13,24 +13,22 @@ module Neuron.CLI.Query
 where
 
 import Colog (WithLog)
-import qualified Data.Aeson.Text as Aeson
-import qualified Data.Map.Strict as Map
+import Data.Aeson (ToJSON)
+import qualified Data.Aeson.Encode.Pretty as AesonPretty
 import qualified Data.Set as Set
-import Data.Some (Some (Some), withSome)
+import Data.Some (withSome)
 import qualified Data.TagTree as TagTree
 import Neuron.CLI.Logging (Message)
 import Neuron.CLI.Types
 import qualified Neuron.Cache as Cache
 import qualified Neuron.Cache.Type as Cache
-import qualified Neuron.Config.Type as Config
 import qualified Neuron.Plugin.Plugins.Tags as Tags
 import qualified Neuron.Reactor as Reactor
 import qualified Neuron.Zettelkasten.Graph as G
 import qualified Neuron.Zettelkasten.Query as Q
-import Neuron.Zettelkasten.Zettel (PluginZettelData (Tags))
 import Relude
 
-runQuery :: (MonadApp m, MonadFail m, MonadApp m, MonadIO m, WithLog env Message m) => QueryCommand -> m ()
+runQuery :: forall m env. (MonadApp m, MonadFail m, MonadApp m, MonadIO m, WithLog env Message m) => QueryCommand -> m ()
 runQuery QueryCommand {..} = do
   Cache.NeuronCache {..} <-
     fmap Cache.stripCache $
@@ -42,28 +40,29 @@ runQuery QueryCommand {..} = do
             Right (ch, _, _) -> pure ch
   case query of
     CliQuery_ById zid -> do
-      let result = G.getZettel zid _neuronCache_graph
-      putLTextLn $ Aeson.encodeToLazyText result
+      let result = G.getZettel zid neuroncacheGraph
+      printJson result
     CliQuery_Zettels -> do
-      let result = G.getZettels _neuronCache_graph
-      putLTextLn $ Aeson.encodeToLazyText result
+      let result = G.getZettels neuroncacheGraph
+      printJson result
     CliQuery_Tags -> do
-      let plugins = Config.getPlugins _neuronCache_config
-      if Some Tags `Map.member` plugins
-        then do
-          let result = Set.unions $ Tags.getZettelTags <$> G.getZettels _neuronCache_graph
-          putLTextLn $ Aeson.encodeToLazyText result
-        else fail "tags plugin is not enabled in neuron.dhall"
+      let result = Set.unions $ Tags.getZettelTags <$> G.getZettels neuroncacheGraph
+      printJson result
     CliQuery_ByTag tag -> do
-      let plugins = Config.getPlugins _neuronCache_config
-      if Some Tags `Map.member` plugins
-        then do
-          let q = TagTree.mkDefaultTagQuery $ one $ TagTree.mkTagPatternFromTag tag
-              zs = G.getZettels _neuronCache_graph
-              result = Tags.zettelsByTag Tags.getZettelTags zs q
-          putLTextLn $ Aeson.encodeToLazyText result
-        else fail "tags plugin is not enabled in neuron.dhall"
+      let q = TagTree.mkDefaultTagQuery $ one $ TagTree.mkTagPatternFromTag tag
+          zs = G.getZettels neuroncacheGraph
+          result = Tags.zettelsByTag Tags.getZettelTags zs q
+      printJson result
     CliQuery_Graph someQ ->
       withSome someQ $ \q -> do
-        result <- either (fail . show) pure $ Q.runGraphQuery _neuronCache_graph q
-        putLTextLn $ Aeson.encodeToLazyText $ Q.graphQueryResultJson q result _neuronCache_errors
+        result <- either (fail . show) pure $ Q.runGraphQuery neuroncacheGraph q
+        printJson $ Q.graphQueryResultJson q result neuroncacheErrors
+  where
+    printJson :: ToJSON a => a -> m ()
+    printJson =
+      putTextLn . decodeUtf8
+        . AesonPretty.encodePretty'
+          AesonPretty.defConfig
+            { -- Sort hash map by keys for consistency
+              AesonPretty.confCompare = compare
+            }