+++ neuron/src/lib/Neuron/Plugin/Plugins/DirTree.hs
@@ -24,16 +24,15 @@ import qualified Data.Set as Set
 import Data.TagTree (Tag, TagNode (..), mkDefaultTagQuery)
 import qualified Data.TagTree as Tag
 import qualified Data.Text as T
-import Data.YAML ((.:?))
-import qualified Data.YAML as Y
 import Neuron.Frontend.Route (NeuronWebT)
 import Neuron.Frontend.Route.Data.Types (DirZettelVal (..))
 import Neuron.Frontend.Widget
   ( ListItem (ListItem_File, ListItem_Folder),
     listItem,
   )
-import qualified Neuron.Markdown as M
+import Neuron.Markdown (lookupZettelMeta)
 import qualified Neuron.Plugin.Plugins.Links as Links
+import Neuron.Plugin.Plugins.Tags (appendTags)
 import qualified Neuron.Plugin.Plugins.Tags as Tags
 import Neuron.Plugin.Type (Plugin (..))
 import Neuron.Zettelkasten.Connection (Connection (Folgezettel, OrdinaryConnection), ContextualConnection)
@@ -44,7 +43,7 @@ import Neuron.Zettelkasten.Resolver (ZIDRef (..))
 import qualified Neuron.Zettelkasten.Resolver as R
 import Neuron.Zettelkasten.Zettel
 import Reflex.Dom.Core
-import Relude hiding (trace, traceShow, traceShowId)
+import Relude
 import qualified System.Directory.Contents as DC
 import System.FilePath (takeDirectory, takeFileName)
 import qualified Text.Pandoc.Definition as P
@@ -55,7 +54,7 @@ plugin :: Plugin DirZettelVal
 plugin =
   def
     { _plugin_afterZettelRead = injectDirectoryZettels,
-      _plugin_afterZettelParse = \myaml -> bimap (afterZettelParse myaml) (afterZettelParse myaml),
+      _plugin_afterZettelParse = bimap afterZettelParse afterZettelParse,
       _plugin_graphConnections = queryConnections,
       _plugin_renderPanel = const renderPanel
     }
@@ -65,17 +64,12 @@ queryConnections ::
   MonadReader [Zettel] m =>
   Zettel ->
   m [(ContextualConnection, Zettel)]
-queryConnections Zettel {..} = do
+queryConnections z = do
   zs <- ask
-  case DMap.lookup DirTree zettelPluginData of
-    Just (Identity (DirZettel _ _ (Just childTag) _)) -> do
+  case lookupPluginData DirTree z of
+    Just (DirZettel _ _ (Just childTag) _) -> do
       pure $ getChildZettels zs childTag
     _ -> pure mempty
-
-metaParser :: Y.Node Y.Pos -> Y.Parser (Maybe DirTreeMeta)
-metaParser yaml = do
-  flip (Y.withMap @(Maybe DirTreeMeta) "DirTree") yaml $ \m -> do
-    m .:? "dirtree"
 
 routePluginData :: ZettelGraph -> DirZettel -> DirZettelVal
 routePluginData g (DirZettel _ parent mChildTag mMeta) =
@@ -94,9 +88,9 @@ getChildZettels zs t =
    in ((conn, ctx),) <$> Tags.zettelsByTag getZettelDirTags zs childrenQuery
 
 getZettelDirTags :: ZettelT c -> Set Tag
-getZettelDirTags Zettel {..} =
+getZettelDirTags z =
   fromMaybe Set.empty $ do
-    _dirZettel_tags . runIdentity <$> DMap.lookup DirTree zettelPluginData
+    _dirZettel_tags <$> lookupPluginData DirTree z
 
 renderPanel :: (DomBuilder t m, PostBuild t m) => DirZettelVal -> NeuronWebT t m ()
 renderPanel DirZettelVal {..} = do
@@ -114,16 +108,12 @@ renderPanel DirZettelVal {..} = do
   where
     pluginDoc = "https://neuron.zettel.page/dirtree"
 
-afterZettelParse :: forall c. Maybe (Y.Node Y.Pos) -> ZettelT c -> ZettelT c
-afterZettelParse myaml z =
-  case runIdentity <$> DMap.lookup DirTree (zettelPluginData z) of
+afterZettelParse :: forall c. ZettelT c -> ZettelT c
+afterZettelParse z =
+  case lookupPluginData DirTree z of
     Just (DirZettel tags mparent (Just childTag) Nothing) ->
-      -- Parse YAML
-      let meta = fromRight Nothing $ case myaml of
-            Nothing -> pure Nothing
-            Just yaml -> M.runYamlParser (metaParser yaml)
-          pluginData = DirZettel tags mparent (Just childTag) meta
-       in z {zettelPluginData = DMap.insert DirTree (Identity pluginData) (zettelPluginData z)}
+      let pluginData = DirZettel tags mparent (Just childTag) getMeta
+       in setDirTreePluginData pluginData
     Just (DirZettel _ _ (Just _) _) ->
       -- This is a directory zettel; nothing to modify.
       z
@@ -136,16 +126,14 @@ afterZettelParse myaml z =
           mparent = do
             guard $ zettelID z /= indexZid
             pure $ parentZettelIDFromPath (zettelPath z)
-          meta = fromRight Nothing $ case myaml of
-            Nothing -> pure Nothing
-            Just yaml -> M.runYamlParser (metaParser yaml)
-       in z
-            { zettelPluginData =
-                DMap.insert
-                  DirTree
-                  (Identity $ DirZettel tags mparent Nothing meta)
-                  (zettelPluginData z)
-            }
+          pluginData = DirZettel tags mparent Nothing getMeta
+       in setDirTreePluginData pluginData
+  where
+    setDirTreePluginData pluginData =
+      -- Add plugin-specific tags to metadata, so the user may query based on them.
+      setPluginData DirTree pluginData $ appendTags (_dirZettel_tags pluginData) z
+    getMeta =
+      lookupZettelMeta @DirTreeMeta "dirtree" (zettelMeta z)
 
 parentZettelIDFromPath :: FilePath -> ZettelID
 parentZettelIDFromPath p =
@@ -223,9 +211,9 @@ rootTag :: TagNode
 rootTag = TagNode "root"
 
 isDirectoryZettel :: ZettelT c -> Bool
-isDirectoryZettel (zettelPluginData -> pluginData) =
-  case DMap.lookup DirTree pluginData of
-    Just (Identity (DirZettel _ _ (Just _childTag) _)) ->
+isDirectoryZettel z =
+  case lookupPluginData DirTree z of
+    Just (DirZettel _ _ (Just _childTag) _) ->
       True
     _ ->
       False