+++ neuron/src/lib/Neuron/Plugin/Plugins/UpTree.hs
@@ -10,7 +10,6 @@
 
 module Neuron.Plugin.Plugins.UpTree (plugin, routePluginData, render) where
 
-import qualified Data.Dependent.Map as DMap
 import Data.Tree (Forest)
 import Neuron.Frontend.Route (NeuronWebT)
 import qualified Neuron.Frontend.Widget.InvertedTree as IT
@@ -25,15 +24,15 @@ import Relude hiding (trace, traceShow, traceShowId)
 plugin :: Plugin (Forest Zettel)
 plugin =
   def
-    { _plugin_afterZettelParse = const $ bimap enable enable,
+    { _plugin_afterZettelParse = bimap enable enable,
       _plugin_routeData = routePluginData,
       _plugin_renderPanel = const render,
       _plugin_css = const IT.style
     }
 
 enable :: ZettelT c -> ZettelT c
-enable z =
-  z {zettelPluginData = DMap.insert UpTree (Identity ()) (zettelPluginData z)}
+enable =
+  setPluginData UpTree ()
 
 routePluginData :: ZettelGraph -> ZettelC -> Forest Zettel
 routePluginData g (sansContent -> z) =