+++ neuron/src/lib/Neuron/Zettelkasten/Zettel/Parser.hs
@@ -1,4 +1,7 @@
+{-# LANGUAGE DeriveAnyClass #-}
+{-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE GADTs #-}
 {-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE OverloadedStrings #-}
@@ -10,13 +13,14 @@
 
 module Neuron.Zettelkasten.Zettel.Parser where
 
+import Data.Default (Default (def))
 import Data.Dependent.Map (DMap)
 import qualified Data.Text as T
-import qualified Data.YAML as Y
-import Neuron.Markdown
+import Data.YAML.ToJSON ()
+import Neuron.Frontend.Theme (titleH1Id)
+import Neuron.Markdown (ZettelParser, lookupZettelMeta)
 import Neuron.Zettelkasten.ID (Slug, ZettelID (unZettelID))
 import Neuron.Zettelkasten.Zettel
-import qualified Neuron.Zettelkasten.Zettel.Meta as Meta
 import Relude
 import qualified Text.Pandoc.Util as P
 
@@ -26,25 +30,19 @@ parseZettel ::
   ZettelID ->
   Text ->
   DMap PluginZettelData Identity ->
-  (Maybe (Y.Node Y.Pos), ZettelC)
+  ZettelC
 parseZettel parser fn zid s pluginData =
   either unparseableZettel id $ do
-    (yamlNode, doc) <- parser fn s
-    meta :: Maybe Meta.Meta <- parseYamlNode @Meta.Meta `traverse` yamlNode
-    let -- Determine zettel title
-        (title, titleInBody) = case Meta.title =<< meta of
-          Just tit -> (tit, False)
-          Nothing -> fromMaybe (unZettelID zid, False) $ do
-            (,True) . P.plainify . snd <$> P.getH1 doc
-        -- Determine other metadata
-        date = Meta.date =<< meta
-        slug = fromMaybe (mkDefaultSlug $ unZettelID zid) $ Meta.slug =<< meta
-        unlisted = Just True == (Meta.unlisted =<< meta)
-    pure $ (yamlNode,) $ Right $ Zettel zid slug fn title titleInBody date unlisted doc pluginData
+    (metadata, doc) <- parser fn s
+    let (tit, doc') = consolidateTitle metadata doc
+        -- Compute final values from user (and post-plugin) metadata
+        slug = fromMaybe (mkDefaultSlug $ unZettelID zid) $ lookupZettelMeta "slug" metadata
+        date = lookupZettelMeta "date" metadata
+    pure $ Right $ Zettel zid metadata slug date fn tit doc' (Just pluginData)
   where
     unparseableZettel err =
       let slug = mkDefaultSlug $ unZettelID zid
-       in (Nothing,) $ Left $ Zettel zid slug fn "Unknown" False Nothing False (s, err) pluginData
+       in Left $ Zettel zid def slug Nothing fn "Unknown" (s, err) (Just pluginData)
     -- We keep the default slug as close to zettel ID is possible. Spaces (and
     -- colons) are replaced with underscore for legibility.
     mkDefaultSlug :: Text -> Slug
@@ -53,12 +51,19 @@ parseZettel parser fn zid s pluginData =
     charsDisallowedInURL :: [Text]
     charsDisallowedInURL =
       [":"]
+    -- Determine the effective title of the zettel, ensuring that the same exists in the document.
+    consolidateTitle metadata doc =
+      case lookupZettelMeta "title" metadata of
+        Just tit ->
+          (tit, P.setTitleH1 titleH1Id tit doc)
+        Nothing ->
+          P.ensureTitleH1 titleH1Id (unZettelID zid) doc
 
 -- | Like `parseZettel` but operates on multiple files.
 parseZettels ::
   ZettelParser ->
   [(ZettelID, (FilePath, (Text, DMap PluginZettelData Identity)))] ->
-  [(Maybe (Y.Node Y.Pos), ZettelC)]
+  [ZettelC]
 parseZettels p files =
   flip fmap files $ \(zid, (path, (s, pluginData))) ->
     parseZettel p path zid s pluginData