+++ neuron/src/lib/Text/Pandoc/Util.hs
@@ -9,7 +9,8 @@
 -- TODO Review and delete what's unused in this module
 module Text.Pandoc.Util
   ( getFirstParagraphText,
-    getH1,
+    setTitleH1,
+    ensureTitleH1,
     plainify,
     PandocLink (..),
     mkPandocAutoLink,
@@ -61,12 +62,24 @@ getFirstParagraphText = listToMaybe . W.query go
       _ ->
         []
 
-getH1 :: Pandoc -> Maybe (B.Attr, [B.Inline])
-getH1 = listToMaybe . W.query go
-  where
-    go :: B.Block -> [(B.Attr, [B.Inline])]
-    go = \case
-      B.Header 1 attr inlines ->
-        [(attr, inlines)]
-      _ ->
-        []
+-- | Ensure that a title H1 is present in the Pandoc doc.
+--
+-- A "title H1" is a H1 that appears as the very first block element of the AST.
+-- If this element is absence, then create it using the given text inline.
+--
+-- Either way, apply the given `id` attribute to the title H1 (so as to
+-- distinguish it from other headings). And return the title plainify'ied.
+ensureTitleH1 :: Text -> Text -> Pandoc -> (Text, Pandoc)
+ensureTitleH1 idAttr defaultTitle = \case
+  (Pandoc meta (B.Header 1 (_id, classes, kw) inlines : rest)) ->
+    -- Add `idAttr` to existing H1
+    let h1 = B.Header 1 (idAttr, classes, kw) inlines
+        tit = plainify inlines
+     in (tit,) $ Pandoc meta (h1 : rest)
+  doc ->
+    (defaultTitle,) $ setTitleH1 idAttr defaultTitle doc
+
+setTitleH1 :: Text -> Text -> Pandoc -> Pandoc
+setTitleH1 idAttr s (Pandoc meta rest) =
+  let h1 = B.Header 1 (idAttr, mempty, mempty) [B.Str s]
+   in Pandoc meta (h1 : rest)