+++ neuron/test/Neuron/Zettelkasten/ZettelSpec.hs
@@ -1,4 +1,3 @@
-{-# LANGUAGE DataKinds #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE NoImplicitPrelude #-}
@@ -8,7 +7,7 @@ module Neuron.Zettelkasten.ZettelSpec
   )
 where
 
-import Data.Tagged (Tagged (Tagged))
+import Data.Default
 import Data.Time.Calendar (fromGregorian)
 import Data.Time.DateMayTime (DateMayTime, mkDateMayTime)
 import Data.Time.LocalTime
@@ -27,7 +26,7 @@ import Test.Hspec
 
 spec :: Spec
 spec = do
-  let noContent :: MetadataOnly = Tagged Nothing
+  let noContent :: MetadataOnly = Nothing
   describe "sortZettelsReverseChronological" $ do
     let mkDay = fromGregorian 2020 3
         mkZettelDay n =
@@ -39,13 +38,12 @@ spec = do
         mkZettel s datetime =
           Zettel
             (ZettelID s)
+            def
             s
+            datetime
             "<spec>.md"
             "Some title"
-            False
             -- (Set.fromList [Tag "science", Tag "journal/class"])
-            datetime
-            False
             noContent
             mempty